+++ Leomard/Models/Lemmy/Forms/ModTools/RemoveCommunity.swift
@@ -1,0 +1,13 @@
+//
+//  RemoveCommunity.swift
+//  Leomard
+//
+//  Created by Konrad Figura on 27/07/2023.
+//
+
+import Foundation
+
+struct RemoveCommunity: Codable {
+    let communityId: Int
+    let removed: Bool
+}