+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## 0.4 (TBA)
+
+## Added
+
+- Mod Tools:
+    - Remove posts
+    - Lock posts
+    - Distinguish comment
+    - Remove comments
+    - Remove community
+    - Edit community
+- Post-locked indicator
+- Distinguished comment indicator
+- Images to Login View
+- Instance icons on the login screen
+- Search for instance on the login screen
+
+### Bug Fixes
+
+- Fixed opening post, if you opened a community from another post
+- Fixed login to some instances caused to faulty decoding of site metadata
+- Fixed displaying videos, if they are added in "URL" part of the post
+
 ## 0.3 (2023-07-27)
 
 ### Added